# Conflicts:
#	Frontend.DB.EF/Models/GeneralContext.cs

+++ Frontend.DB.EF/Models/UserArc.cs
@@ -17,7 +17,7 @@ namespace Frontend.DB.EF.Models
         public string IdImageA { get; set; }
         public string IdImageB { get; set; }
         public string IdImageC { get; set; }
-        public byte? KycStatus { get; set; }
+        public short? KycStatus { get; set; }
         public DateTime? KycStatusUpdateTime { get; set; }
         public DateTime? CreateTime { get; set; }
         public DateTime? UpdateTime { get; set; }
